fix: use non-obsolete EntityId.ToULong/FromULong in MCPObjectId
The EntityId<->int implicit cast is itself obsolete on Unity 6.5 (CS0619)
— entity ids will no longer be representable by an int. Switch the shim to
expose ids as long and round-trip through EntityId.ToULong / EntityId.FromULong
(both non-obsolete). long keeps existing int call sites working via implicit
widening; the one HashSet<int> dedup site in MCPUMACommands is widened to
HashSet<long>.

Verified: 0 errors / 0 warnings on Unity 6000.2 (legacy branch).

+++ Editor/MCPUMACommands.cs
@@ -1897,7 +1897,7 @@ namespace UnityMCP.Editor
         {
             var textRecipes = indexer.GetAllAssets<UMATextRecipe>();
             var wardrobeRecipes = indexer.GetAllAssets<UMAWardrobeRecipe>();
-            var seen = new HashSet<int>();
+            var seen = new HashSet<long>();
             var combined = new List<UMATextRecipe>();
             if (textRecipes != null)
                 foreach (var r in textRecipes)
